Updated INIT to avoid more filename issues.
If a user included “R1” in the name of a raw reads file then INIT would
get confused if that meant the the file contained the first reads or if
R1 was just part of the file name. Now we look for _R1_ and _R2_ to
determine if the file contains the first or second reads.