enhancements, bug fixes.
* renamed RUMALIGN to be RUM
* changed RUM output to be consistent with STAR output format
(sampleID.rum.sorted.bam and sampleID.rum.unique.bam).
* added error checking to RUM.
* renamed rum.trim to rum
* more specific error reporting for all modules.
* major revamp of STATS module allowing for more user control and
better incorporation of future modules. This change permeates all
modules that might include stats. Users can now specify which modules
to use for stats and the ordering of the modules. Modules are now
expected to contain a function that returns the stats for that module.